v0.0.32 Fixed Boz and Beetle bugs, replaced RockChooser with more flexible ProbabilityTable script

+++ Assets/Scripts/Level Layout/Formation System/FormationProfile.cs
@@ -8,8 +8,8 @@ public class FormationProfile {
 
 	public string name;
 
-	public int minimumDistance;
-	public int maximumDistance;
+	public float minimumDistance;
+	public float maximumDistance;
 
 	public float probabilityWeight;
